# storage_proxy: reduce allocations in send_to_local_endpoints()