@gleb-cloudius 2025-07-31
> > The instruction count reduction is a not very impressive 70/op:
> > before
> > ```
> > instructions_per_op:
> > 	mean=   53412.22 standard-deviation=32.12
> > 	median= 53420.53 median-absolute-deviation=20.32
> > 	maximum=53462.23 minimum=53290.06
> > ```
> > 
> > 
> >     
> >       
> >     
> > 
> >       
> >     
> > 
> >     
> >   
> > after
> > ```
> > instructions_per_op:
> > 	mean=   53350.32 standard-deviation=32.38
> > 	median= 53353.71 median-absolute-deviation=13.60
> > 	maximum=53415.20 minimum=53222.24
> > ```
> 
> Are those numbers with PGO or without? And if they are without, are they truly representative?

The change make sense regardless of numbers.