@xemul 2025-07-31
> send_to_local_endpoints() computes sets of endpoints to which we send mutation ...

send_to_<i>live</i>_endpoints ?