@avikivity 2025-07-31
v2:
  - re-applied crucial change to small_vector which got lost while measuring performance
  - corrected function name in changelog