@gleb-cloudius 2025-07-31
> btw, the entire data structure should be copied from effective_replication_map instead of being computed on the fly.

effective_replication_map neither knows node liveness not it takes a type of write into account (read repair writes are handled differently).