@avikivity 2025-07-31
> > btw, the entire data structure should be copied from effective_replication_map instead of being computed on the fly.
> 
> effective_replication_map neither knows node liveness not it takes a type of write into account (read repair writes are handled differently).

Dead nodes could be filtered out.

```c++

struct replication_plan {
   utils::small_vector<host_id, 9> all_nodes;
   utils::small_vector<std::span<uint16_t>, 5> forwarding_groups; // pointing into all_nodes
};
```

This is enough to cover 3-DC replication with no allocations, and can be recomputed on every topology change.

Read repair can be computed on the fly.