@gleb-cloudius 2025-07-31
> > > btw, the entire data structure should be copied from effective_replication_map instead of being computed on the fly.
> > 
> > 
> > effective_replication_map neither knows node liveness not it takes a type of write into account (read repair writes are handled differently).
> 
> Dead nodes could be filtered out.

And will require allocations. Do not forget that we actually save live endpoints in the write handler to track replies, so this code goes after filtering already happened.

> 
> ```c++
> struct replication_plan {
>    utils::small_vector<host_id, 9> all_nodes;
>    utils::small_vector<std::span<uint16_t>, 5> forwarding_groups; // pointing into all_nodes
> };
> ```
> 
> This is enough to cover 3-DC replication with no allocations, and can be recomputed on every topology change.
> 
> Read repair can be computed on the fly.

Yeah. Lets have 3 code paths (regular, dead endpoints, repair) instead of one to save exactly what before going to the slowest possible code anyway (networking).

Did you see this code shows in profiling? The change makes sense because it removes a complication that was needed by already re-written code, not because it has a huge impact on performance.