@avikivity 2025-08-01
> #### 🔴 CI State: FAILURE
> Mode	Stage	Status	Node
> [Framework test](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/18679/)	✅	spider9.cloudius-systems.com
> dev	[Build](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/18679/)	✅	spider7.cloudius-systems.com
> [test.py](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/18679/)	❌	spider7.cloudius-systems.com
> release	[Build](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/18679/)	❌	spider4.cloudius-systems.com
> debug	[Build](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/18679/)	✅	spider6.cloudius-systems.com
> [test.py](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/18679/)	❌	spider6.cloudius-systems.com
> #### Build Details:
> * Duration: 1 hr 13 min

https://github.com/scylladb/scylla-pkg/issues/5367