@yaronkaikov 2025-08-03
> > #### 🔴 CI State: FAILURE
> > Mode	Stage	Status	Node
> > [Framework test](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/18721/)	✅	spider11.cloudius-systems.com
> > dev	[Build](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/18721/)	❌	spider8.cloudius-systems.com
> > release	[Build](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/18721/)	❌	spider8.cloudius-systems.com
> > debug	[Build](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/18721/)	❌	spider8.cloudius-systems.com
> > #### Build Details:
> > 
> > * Duration: 6 hr 57 min
> 
> ```
> 05:32:45  + podman login docker.io --username **** --password ****
> 05:32:45  time="2025-08-02T05:32:44+03:00" level=warning msg="RunRoot is pointing to a path (/run/user/1000/containers) which is not writable. Most likely podman will fail."
> 05:32:45  Error: default OCI runtime "crun" not found: invalid argument
> ```
> 
> @yaronkaikov

I know i know, working to solve it